+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  - handle field `load_as_table` in style `entryList`
  - add `url_cancel` to `formUserInput`
  - add `confirmation message` functionality to style `button`
+ - add `confirmation message` functionality to style `formUserInput`
 
 ### Bugfix
  - #2 - fix the validation, `url` is correctly recognized with the capacitor.